Feat: Map 컴포넌트 분리
